support negative numbers

+++ src/Lucene.Net.Core/Codecs/Lucene40/Lucene40DocValuesReader.cs
@@ -241,7 +241,7 @@ namespace Lucene.Net.Codecs.Lucene40
 
             public override long Get(int docID)
             {
-                return Values[docID];
+                return (sbyte)Values[docID];
             }
         }
